refactor(examples): fix import paths in moved modules
